+++ apps/web/src/entities/exercisePreset/api/useCreateExercisePreset.ts
@@ -1,7 +1,5 @@
-import type { MockedResponse } from '@apollo/client/testing';
 import { useMutation } from '@apollo/client'
 import CreateExercisePresetGql from './graphql/mutation/CreateExercisePresetGql';
-import { ExercisePresetMockData } from '.'
 
 export default function useCreateExercisePreset() {
   return useMutation<CreateExercisePresetResponse, CreateExercisePresetVariable>(CreateExercisePresetGql, {
@@ -19,26 +17,4 @@ export default function useCreateExercisePreset() {
       })
     }
   })
-}
-export const CreateExercisePresetMock: MockedResponse<
-  CreateExercisePresetResponse,
-  CreateExercisePresetVariable
-> = {
-  request: {
-    query: CreateExercisePresetGql,
-  },
-  result: (v) => {
-    const id = Math.max(...Object.keys(ExercisePresetMockData).map(Number)) + 1
-    ExercisePresetMockData[id] = {
-      name: v.exercisePreset.name,
-      id: id,
-      deps: 0,
-      __typename: 'ExercisePreset'
-    }
-    return {
-      data: {
-        createExercisePreset: ExercisePresetMockData[id]
-      }
-    }
-  }
 }
