+++ apps/web/src/entities/exercisePreset/api/useGetExercisePresetWithListByOffset.mock.ts
@@ -1,0 +1,26 @@
+import type { MockedResponse } from '@apollo/client/testing';
+import GetExercisePresetWithListByOffset from './graphql/query/GetExercisePresetWithListByOffset';
+import { ExercisePresetMockData } from '.'
+
+const GetExercisePresetWithListListMock: MockedResponse<
+  GetExercisePresetWithListByOffsetResponse,
+  GetExercisePresetWithListByOffsetVariable
+> = {
+  request: {
+    query: GetExercisePresetWithListByOffset
+  },
+  result: (v) => {
+    return {
+      data: {
+        getExercisePresetWithListByOffset: Object
+          .values(ExercisePresetMockData)
+          .splice(v.offset, v.size)
+          .map(v => {
+            v.__typename = 'ExercisePresetWithList'
+            return v
+          })
+      }
+    }
+  }
+}
+export default GetExercisePresetWithListListMock
